Report_1976_1986
This is the Economic Report to the Governor published in 1989. The data extracted form this file was used to account for the time period between 1976-1989. The file was too big, so I could not upload it here. But you can download it from this webpage: https://jp.pr.gov/apendice-estadistico-del-informe-economico-a-la-gobernador/